switch to react-table for e2e purposes
I couldn't get the blueprintjs rows to respond to clicks, even from the debugger. So I've switched to the folderlist being a react-table, and will do the same with the file list in the future. Already knew the blueprintjs wasn't going to be good for all the tables, because it apparently cannot host editable elements.