Fix a few e2e tests
The main thing is that you currently need to run ``yarn build-production`` first. The other problem was that the app title had changed from sayless to sayless mac. A button id had changed. Finally, because we're currently having trouble driving menus, runner.restart() had to change the new process.env flag that controls whether we start in the start screen or not, so that the tests expecting us to restart into our new project would get that behavior.