simplify date handling
Now we just try to quickly get the date into a yyyy-mm-dd format, and don't mess with it after that.

Had to change tsconfig settings and how commonjs modules are imported in order to get both jest and the app to both be happy with moment. See https://github.com/aurelia/skeleton-navigation/issues/606