Remove entitlements
With MacOS entitlements specified, SMx crashes immediately, as something it is doing isn't allowed by the listed entitlements. E.g., maybe reading the last-opened project. I didn't see a way to enable that, but if you just don't provide this file at all, it's fine, at least as of Mojave.