Fix the return of the undead meta files
Steps:
Drag an image into a person.
In SayMorex, Click on the person.
In explorer, see that the file and its meta file are there.
In SayMorex, delete it.
In explorer, see that both files are removed.
In SayMorex, Click on the person.
In explorer, see that the meta file came back.