Let view of IMDI of resource files access the props they inherit from the folder
E.g., we have access on the folder, and want to see how that comes out in an IMDI of an image in that folder. It was already going to come out correctly in an actual export, but now it comes out correct in the "IMDI" tab of all the files.