WIP upgrading to mobx 6
I just did this as a matter of maintenance, so that it's easier to find the right documentation on mobx stuff.

They've deprecated @observable.
As for @observer, the next step is to wrap them in `const RoleChooser = observer(class roleChooser ...`. You can see that I did this on RoleChooser.tsx. Not hard, but there are 25 classes to do!

There are other problems showing up while testing. Decided that this is not the right time for this change.