Fix 238, No <Age> when Session/Date is Unspecified
If the Session/Date is not given in Lameta, the IMDI value is Unspecified, if a Contributor has a BirthDate, there is no Age element and IMDI not valid. The result shoudl rather be the same as when a Contributor has BirthDate Unspecified, i.e. Age is Unspecified and a comment “Could not compute Age”. The comment “The following is based on today’s date” seems not to be true in either case(?) and this is probably better, since it’s irrelevant how old the speaker is today when we’re analysing a conversation they took part in 20 years ago.