LAM-71 Person without info - omit description field for unresolved contributors
https://linear.app/lameta/issue/LAM-71/person-without-info

Previously when a Person entity was created for an unresolved contributor (no matching person record found), we added a description field with the message 'The lameta project could not find a matching Person for this contributor, so we do not have any further information'.

This commit removes that description field entirely. When we don't have information about a person, we simply omit the description rather than adding a placeholder message.

Changes:
- RoCratePeople.ts: Removed the line that sets the description for unresolved contributors
- RoCrateExporter.unresolved-contributor.spec.ts: Updated test expectations to verify description is undefined

Tests: All RO-Crate tests pass (yarn test:rocrate)